rename global-web3 to configure-web3
