+++ configure.js
@@ -1,4 +1,4 @@
-const { setWeb3 } = require('./src/global-web3');
+const { setWeb3 } = require('./src/configure-web3');
 
 let loadedConfig;
