+++ src/setup.js
@@ -1,6 +1,6 @@
 require('../configure')();
 
-const web3 = require('./global-web3').getWeb3();
+const web3 = require('./configure-web3').getWeb3();
 
 const chai = require('chai');
 const BN = web3.utils.BN;
